fix(cluster) store and use protocol_version from first coordinator
Not doing so would be harmful when a connection would be kept alive, but
when being reused, `connect()` is called but doesn't check if the
procotol is supported (only does so when `getreusedtimes() == 0`). We
now store the protocol version in the cluster's `peers_opts` table.

All nodes from the cluster must currently use the same protocol version.