feat(cluster) better keyspace pooling
Taking advantage of the recent fix in the init module that pools sockets
by their keyspace, this avoids setting a keyspace before each operation
in order to allow for `coordinator_options` in `execute()` and
`batch()`.

Instead, we set coordinators their keyspace when opening their
connection, as it should be. Cassandra keyspaces are expected to exist
when we connect to it via a cluster instance, or else the connection
will fail, as it should be. If no keyspace is set for the cluster, none
will be used. If a keyspace is set at the `coordinator_options` argument
(per-query level of granularity), then the cluster will set the keyspace
beore spawning the coordinator, hence it will always use a different
connection pool.