Fix 01-rockspec tests
These tests assume the 'find *.rockspec' command will return the 2
rockspec files in a specific order (first the prod one and then dev).

This is not always true. It might depend on the OS, find implementation
or something else, I don't know. But we can't assume any order.

So I've changed the code so that it'll now try to match the filenames.